Fix auth bugs 🤩
