+++ frontend/src/components/Header.js
@@ -3,7 +3,7 @@ import "../App.css";
 import { classNames } from "@hkamran/utility-web";
 import Dropdown from "react-bootstrap/Dropdown";
 import GenreDDI from "./GenreDDI";
-
+import { useNavigate } from "react-router-dom";
 
 export default function Header({
     setSearchQuery,
@@ -15,9 +15,19 @@ export default function Header({
     genreFilter,
 }) {
     let leftElement, rightElement, dropdown;
+    const navigate = useNavigate();
 
     if (useSearchBar) {
-        leftElement = <a href={href}>{hrefName}</a>;
+        leftElement = (
+            <a
+                onClick={() => {
+                    navigate(href);
+                }}
+                className="Link"
+            >
+                {hrefName}
+            </a>
+        );
         rightElement = (
             <input
                 type="text"
@@ -41,28 +51,28 @@ export default function Header({
         );
     } else {
         leftElement = (
-            <a href={"https://thechampionproject.org/"}>Our mission</a>
+            <a href = "https://thechampionproject.org/">
+                Our mission
+            </a>
         );
         rightElement = (
             <a
-                href={"/login"}
+                onClick={() => navigate("/login")}
                 style={{ position: "absolute", top: ".5em", right: "1em" }}
+                className="Link"
             >
                 Sign in/up
             </a>
         );
         dropdown = <></>;
-
     }
     return (
         <>
             <div className="fixed-top navbar NavHead">
                 {leftElement}
-
                 {dropdown}
                 <h3 className="CPStyleMobile">TCP</h3>
                 <h3 className="CPStyleFull">The Champion Project</h3>{" "}
-
                 {rightElement}
             </div>
