+++ frontend/src/components/Login.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import {
     Grid,
     Paper,
@@ -12,7 +12,7 @@ import {
     DialogContent,
 } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-import { sendPasswordReset, signInAuth, auth } from "../FirebaseFunctions";
+import { resetPasswordAuth, signInAuth, auth } from "../FirebaseFunctions";
 import { useAuthState } from "react-firebase-hooks/auth";
 
 const Login = ({ handleChange }) => {
@@ -52,7 +52,7 @@ const Login = ({ handleChange }) => {
         } else {
             try {
                 setOpen(false);
-                await sendPasswordReset().then(() =>
+                await resetPasswordAuth().then(() =>
                     alert("Success! Check your spam folder for the reset email")
                 );
             } catch (err) {
