+++ frontend/src/components/UserProtection.js
@@ -1,17 +1,16 @@
 import React, { useEffect } from "react";
-import { getSignedInUserNameFB } from "../FirebaseFunctions";
 import { auth } from "../FirebaseFunctions";
 import { useAuthState } from "react-firebase-hooks/auth";
 export default function UserProtection() {
-    const [user] = useAuthState(auth);
+    const [user, loading] = useAuthState(auth);
     useEffect(() => {
         const getUsername = async () => {
-            if (!user) {
+            if (user === null && !loading) {
                 alert("You must be signed in to view this page");
                 window.location = "/login";
             }
         };
         getUsername();
-    }, []);
+    }, [user]);
     return <></>;
 }
