+++ frontend/src/pages/GiftPage.js
@@ -4,7 +4,8 @@ import TableStructure from "../components/TableStructure";
 import Header from "../components/Header";
 import "../App.css";
 import UserProtection from "../components/UserProtection.js";
-
+import { useAuthState } from "react-firebase-hooks/auth"; // use this in every page
+import { auth } from "../FirebaseFunctions"; // use this in every page
 
 export default function ManagePage() {
     const [managedBook, setManagedBook] = useState(null); // The book entry that needs to be edited
@@ -18,10 +19,16 @@ export default function ManagePage() {
     const [book, setBook] = useState(null);
     const [books, setBooks] = useState(null);
     let [genreFilter, setGenreFilter] = useState("All");
+//    const [user, loading] = useAuthState(auth);
+//
+//    useEffect(() => {
+//        if (!loading && !user) {
+//            navigate("/login");
+//        }
+//    }, [user]);
 
     return (
         <>
-
             <UserProtection />
             <Header
                 setSearchQuery={setSearchQuery}
@@ -31,7 +38,6 @@ export default function ManagePage() {
                 useSearchBar={true}
                 setGenreFilter={setGenreFilter}
                 genreFilter={genreFilter}
-
             />
 
             <TableStructure
@@ -45,7 +51,6 @@ export default function ManagePage() {
                 books={books}
                 setBooks={setBooks}
                 genreFilter={genreFilter}
-
             />
             <GiftConfirmation
                 showGC={showGC}
