+++ frontend/src/pages/HomePage.js
@@ -2,13 +2,14 @@ import Button from "react-bootstrap/Button";
 import React, { useState, useEffect } from "react";
 import VolunteerDates from "../components/VolunteerDates";
 import { useNavigate } from "react-router-dom";
+import UserProtection from "../components/UserProtection";
 import {
-    getSignedInUserInfoFB,
     signOutUser,
-    signUpForDate,
+    changeDateFB,
+    auth,
+    getSignedInUserInfoFB,
 } from "../FirebaseFunctions";
 import { useAuthState } from "react-firebase-hooks/auth";
-import { auth } from "../FirebaseFunctions";
 
 export default function HomePage() {
     const [selectedDateIDs, setSelectedDateIDs] = useState([]);
@@ -16,25 +17,28 @@ export default function HomePage() {
     const [availableDates, setAvailableDates] = useState([]);
     const [username, setUsername] = useState("");
     const [fullUserName, setFullUsername] = useState(null);
-    const [user] = useAuthState(auth);
+    const [user, loading] = useAuthState(auth);
 
     const navigate = useNavigate();
 
     useEffect(() => {
         const getUsername = async () => {
-            if (!user) {
+            try {
+                if (!user && !loading) {
+                    alert("You must be signed in to view this page");
+                    window.location.href = "/login";
+                }
+                const info = await getSignedInUserInfoFB(user.uid);
+
+                setUsername(info.name.split(" ")[0]);
+                setFullUsername(info.name);
+            } catch {
                 alert("You must be signed in to view this page");
-                window.location = "/login";
+                window.location.href = "/login";
             }
-
-            const info = await getSignedInUserInfoFB(user.uid);
-
-            setUsername(info.name.split(" ")[0]);
-            setFullUsername(info.name);
         };
         getUsername();
     }, []);
-
 
     const submit = async (e) => {
         e.preventDefault();
@@ -50,7 +54,7 @@ export default function HomePage() {
 
         let request = "Error";
         try {
-            request = await signUpForDate(selectedDateIDs, unselectedDateIDs);
+            request = await changeDateFB(selectedDateIDs, unselectedDateIDs);
         } catch {
             alert("There was an error with your request");
         }
@@ -96,12 +100,12 @@ export default function HomePage() {
 
                 <br />
                 <h2>Chose a page to go to: </h2>
-                <Button variant="primary" href="/manage">
+                <Button variant="primary" onClick={() => navigate("/manage")}>
                     Manage the book database
                 </Button>
                 <br />
                 <br />
-                <Button variant="primary" href="/gift">
+                <Button variant="primary" onClick={() => navigate("/gift")}>
                     Gift books
                 </Button>
                 <br />
@@ -112,16 +116,13 @@ export default function HomePage() {
 
                 <br />
                 <br />
-                <Button variant="primary" href="/admin">
-                    Admin Page
+                <Button variant="primary" onClick={() => navigate("/admin")}>
+                    Change Event Dates
                 </Button>
                 <br />
                 <br />
                 <h2>Or sign up to volunteer: </h2>
-                <h5>
-                    Available Dates (If you're on mobile, you may need to press
-                    aA, Request Desktop Site):
-                </h5>
+                <h5>Available Dates:</h5>
                 <VolunteerDates
                     availableDates={availableDates}
                     setAvailableDates={setAvailableDates}
