+++ frontend/src/pages/ManagePage.js
@@ -7,8 +7,8 @@ import Header from "../components/Header";
 import UserProtection from "../components/UserProtection.js";
 import { useNavigate } from "react-router-dom";
 import { getSignedInUserInfoFB } from "../FirebaseFunctions";
-import { useAuthState } from "react-firebase-hooks/auth";
-import { auth } from "../FirebaseFunctions";
+import { useAuthState } from "react-firebase-hooks/auth"; // use this in every page
+import { auth } from "../FirebaseFunctions"; // use this in every page
 export default function ManagePage() {
     const [book, setBook] = useState(null); // The book that gets passed to popup
     const [showEditPoup, setShowEditPopup] = useState(false); // Show the popup
@@ -24,27 +24,18 @@ export default function ManagePage() {
 
     const [scanMode, setScanMode] = useState(false);
     let [genreFilter, setGenreFilter] = useState("All");
+    //const [user, loading] = useAuthState(auth);
     let navigate = useNavigate();
 
     let handleAddBook = (e) => {
         setShowAddPopup(true);
     };
 
-    useEffect(() => {
-        const getVerification = async () => {
-            const response = await getSignedInUserInfoFB();
-
-            if (response === "No user signed in") {
-                alert("You must be signed in to view this page");
-                navigate("/login");
-            }
-            if (!(response.uploadedForm && response.watchedVideo)) {
-                alert("You must be a verified volunteer to view this page");
-                navigate("/home");
-            }
-            getVerification();
-        };
-    }, []);
+    //useEffect(() => {
+    //    if (!loading && !user) {
+    //        navigate("/login");
+    //    }
+    //}, [user]);
 
     useEffect(() => {
         const statusKeyboardInput = (e) => {
@@ -60,6 +51,7 @@ export default function ManagePage() {
     return (
         <>
             <UserProtection />
+
             <Header
                 setSearchQuery={setSearchQuery}
                 alert={alert}
@@ -100,7 +92,6 @@ export default function ManagePage() {
                 setBook={setBook}
                 setManagedBook={setManagedBook}
                 setAlert={setAlert}
-                
                 scanMode={scanMode}
                 setScanMode={setScanMode}
                 setShowAddPopup={setShowAddPopup}
