--- the-champion-project-database-default-rtdb-export.json
@@ -1,115 +1,0 @@
-{
-    "active": {
-        "005c66b7-089f-4a15-9c88-f5571d0dc483": {
-            "Genre": "Be Inspired",
-            "Inventory": 1,
-            "Needed": 0,
-            "Price": 7.99,
-            "Sub-Genre": "P - Be Inspired/People (Blue)",
-            "Timestamp": "11/5/2022 10:00:58",
-            "Title": "what color is my world?"
-        },
-        "006ccf38-c48a-4f76-80d6-f6b3fd0a8716": {
-            "Genre": "Explore",
-            "Inventory": 1,
-            "Needed": 0,
-            "Price": 4.99,
-            "Sub-Genre": "",
-            "Timestamp": "10/16/2022 16:07:47",
-            "Title": "Press Start! Game Over, Super Rabbit Boy!"
-        },
-        "00716f53-a535-4f6b-8c62-019338b47dc1": {
-            "Genre": "Laugh",
-            "Inventory": 1,
-            "Needed": 0,
-            "Price": 4.99,
-            "Sub-Genre": "",
-            "Timestamp": "10/16/2022 15:50:29",
-            "Title": "Animal Joke Book"
-        }
-    },
-    "archive": {
-        "206fea77-2d60-4705-89664-182f132eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 10.99,
-            "Title": "Blank"
-        },
-        "206fea77-2d60-4964-182f132eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 11.99,
-            "Title": "Blank"
-        },
-        "206fea77-2d61-4705-8964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "20afea77-2d60-4705-8964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "216fea77-2d60-4705-7964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "216fea77-2d60-4705-8964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "246fea77-2d60-4705-8964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "306fea77-2d60-4705-8964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "406fea77-2d60-4705-8964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "906fea77-2d60-4705-8964-182f1328eff18": {
-            "ArchiveDates": ["2022-11-24T09:32:33.778Z"],
-            "Genre": "Laugh",
-            "Price": 7.99,
-            "Title": "Blank"
-        },
-        "d917a51c-a73e-4a30-afa4-56cb6f207dd2": {
-            "ArchiveDates": ["2022-12-03T05:14:09.407Z"],
-            "Genre": "Laugh",
-            "Price": 6.99,
-            "Title": "1, 2, 3… By the sea"
-        }
-    },
-    "volunteer-dates": {
-        "637856178687361": {
-            "startDate": "2022-12-16 12:00:00",
-            "endDate": "2022-12-16 14:00:00",
-            "location": "The Library",
-            "id": "637856178687361",
-            "volunteers": [""]
-        },
-        "895729875758493": {
-            "startDate": "2022-12-14 13:00:00",
-            "endDate": "2022-12-14 16:00:00",
-            "location": "12 West Avenue",
-            "id": "895729875758493",
-            "volunteers": [""]
-        }
-    }
-}
